# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "Petship API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://api.petship.pt",
            "description": "Petship API"
        }
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Utilizadores"
                ],
                "summary": "Login",
                "description": "Realiza o login de um utilizador.",
                "operationId": "72e2dd71e8d5a15cf6308453f61513df",
                "requestBody": {
                    "description": "Dados de login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "example@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "utilizador": {
                                            "$ref": "#/components/schemas/Utilizador"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|SFrEYKGln7vdvwdN09Ai9gBeA15Uu8hk7GcSAgQR"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Email não verificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Email não verificado",
                                            "type": "string",
                                            "example": "Email não verificado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Email ou password incorretos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Email ou password incorretos",
                                            "type": "string",
                                            "example": "Email ou password incorretos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/registar": {
            "post": {
                "tags": [
                    "Utilizadores"
                ],
                "summary": "Registar",
                "description": "Regista um novo utilizador.",
                "operationId": "007a1da2c8bc6130c0ac5b58b217d3db",
                "requestBody": {
                    "description": "Dados de registo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Joao Ferreira"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "example@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "tipo": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registo realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "description": "Registo realizado com sucesso",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Erro de validação",
                                            "type": "string"
                                        },
                                        "errors": {
                                            "description": "Detalhes dos erros de validação",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/associacoes": {
            "get": {
                "tags": [
                    "Utilizadores"
                ],
                "summary": "Listar Associações",
                "description": "Retorna uma lista de associações.",
                "operationId": "a724c60c5400aa3bf904d8e0174b4d72",
                "responses": {
                    "200": {
                        "description": "Lista de associações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "associacoes": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Não autorizado",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}